requirements.txt
fastapi
uvicorn
transformers
torch
pillow
pyngrok
nest_asyncio
python_multipart